debug: Don't clear DEBUG setting based of ctor options and use test runner trace functions if present.
The DEBUG mode is currently just a global flag, but can be influenced by
constructor options for various classes.  Ideally there would be a debug
toggle for each class, but that may be overkill.  This at least now
allows a debug flag set in one class to not be killed when another is
constructed without its debug flag set.  Another option would be to
leverage the NODE_DEBUG style of enabling.